requirements.txt
google-generativeai
django
